Move Dataset queries into TypeScript module & separate graphql schema file Move Knex db singleton into its own file for TS compatibility
