+++ metaspace/graphql/package.json
@@ -13,6 +13,7 @@
   },
   "license": "ISC",
   "dependencies": {
+    "@types/amqplib": "^0.5.9",
     "@types/aws-sdk": "^2.7.0",
     "@types/bcrypt": "^2.0.0",
     "@types/config": "^0.0.34",
