+++ metaspace/graphql/schema.ts
@@ -4,6 +4,7 @@ import * as fs from 'fs';
 
 export const mergedSchemas = mergeTypes([
   fs.readFileSync('schema.graphql', 'utf8'),
+  fs.readFileSync('schemas/dataset.graphql', 'utf8'),
   fs.readFileSync('schemas/user.graphql', 'utf8'),
   fs.readFileSync('schemas/group.graphql', 'utf8'),
   fs.readFileSync('schemas/project.graphql', 'utf8'),
