+++ metaspace/graphql/src/bindingTypes.ts
@@ -33,6 +33,31 @@ export type ProjectSource = ProjectModel & { currentUserRole: UserProjectRole | 
 export type UserProjectSource = {
   [field in keyof UserProjectModel]: field extends 'user' ? UserSource : UserProjectModel[field]
 };
+export interface DatasetSource {
+  scopeRole: ScopeRole;
+  _source: {
+    ds_id: string;
+    ds_name: string;
+    ds_upload_dt: string;
+    ds_config: any;
+    ds_meta: any;
+    ds_status: string;
+    ds_input_path: string;
+    ds_is_public: boolean;
+    ds_mol_dbs: string[];
+    ds_adducts: string[];
+    ds_acq_geometry: any;
+    ds_submitter_id: string;
+    ds_submitter_name: string;
+    ds_submitter_email: string;
+    ds_group_id: string | null;
+    ds_group_name: string | null;
+    ds_group_short_name: string | null;
+    ds_group_approved: boolean;
+    ds_project_ids?: string[];
+    annotation_counts: any[];
+  };
+}
 
 // Utility to extract the type of the `args` field from a query/mutation in binding.ts
 // Usage: customField(source, args: ArgsFromBinding<CustomType['customField']>)
