+++ metaspace/graphql/src/modules/dataset/controller/index.ts
@@ -1,7 +1,13 @@
 import Mutation from './Mutation';
+import Query from './Query';
+import Dataset from './Dataset';
+import Subscription from './Subscription';
 import {IResolvers} from 'graphql-tools';
 import {Context} from '../../../context';
 
 export const Resolvers = {
-  Mutation
+  Mutation,
+  Query,
+  Dataset,
+  Subscription,
 } as IResolvers<any, Context>;
